requirements.txt
altair==6.0.0
annotated-doc==0.0.4
anyio==4.12.1
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astroid==4.0.3
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.14.3
black==26.1.0
bleach==6.3.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==7.0.5
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
chrome-gnome-shell==0.0.0
click==8.3.3
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.19
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
dill==0.4.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fasteners==0.14.1
fastjsonschema==2.21.2
filelock==3.28.0
flake8==7.3.0
fonttools==4.61.1
fqdn==1.5.1
fsspec==2026.3.0
future==0.18.2
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.3
imbalanced-learn==0.14.1
importlib-metadata==4.6.4
ipykernel==7.1.0
ipython==8.38.0
ipywidgets==8.1.8
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
lark==1.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.11.0
lightgbm==4.6.0
llvmlite==0.46.0
lockfile==0.12.2
louis==3.20.0
lxml==6.0.2
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==4.0.0
MarkupSafe==2.0.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.2.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.2
mypy_extensions==1.1.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
notebook==7.5.3
notebook_shim==0.2.4
numba==0.63.1
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.29.2
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
oauthlib==3.2.0
olefile==0.46
opencv-python==4.13.0.90
openpyxl==3.1.5
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.5
pathspec==1.0.4
patsy==1.0.2
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.5.1
plotly==6.5.2
pooch==1.9.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycairo==1.20.1
pycodestyle==2.14.0
pycparser==3.0
pycups==2.0.1
pydeck==0.9.2
pyflakes==3.4.0
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==4.0.4
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.3.2
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.2.0
python-json-logger==4.0.0
python-pptx==1.0.2
pytokens==0.4.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.1.0
referencing==0.37.0
regex==2026.4.4
reportlab==3.6.8
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==2.1.0
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.5
smmap==5.0.3
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
stack-data==0.6.3
statsmodels==0.14.6
streamlit==1.56.0
sympy==1.14.0
systemd-python==234
tenacity==9.1.4
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.4.0
tomlkit==0.14.0
torch==2.11.0+cpu
torchvision==0.26.0+cpu
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.41.2
triton==3.6.0
typer==0.25.0
typing_extensions==4.15.0
tzdata==2025.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.5.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
xdg==5
xgboost==3.1.3
xkit==0.0.0
xlrd==2.0.2
xlsxwriter==3.2.9
zipp==1.0.0
